+++ src/FScript.LanguageServer.CSharp/Program.cs
@@ -1,3 +1,4 @@
-using FScript.LanguageServer;
+using FScript.LanguageServer.CSharp;
 
-LspServer.run();
+var server = new LspServer();
+server.Run();